fix(db): correct the two PostgreSQL type mismatches the profile sweep found
Saving reading preferences was broken on PostgreSQL and nothing noticed.

`typography_preference.distraction_free` is BOOLEAN, and the PostgreSQL
VALUES list bound it bare, so `PATCH /settings/typography` answered 500
with "column \"distraction_free\" is of type boolean but expression is of
type bigint". The write is now `?::int::boolean`, matching the treatment
`contains_spoilers` and `is_public` already had.

`font_scale` and `line_height` were declared REAL. The repository decodes
them as f64, and PostgreSQL's REAL is 4-byte FLOAT4 against Rust's FLOAT8,
so the read that followed a successful write failed with "mismatched
types; Rust type `f64` (as SQL type `FLOAT8`) is not compatible with SQL
type `FLOAT4`". SQLite's REAL is 8-byte, so the two dialects meant
different widths for the same declaration. Both are now DOUBLE PRECISION,
and the read casts them explicitly so the contract is stated rather than
inherited from the column type.

Both were found by enumerating the PostgreSQL schema's boolean and
4-byte-float columns and checking every bind and read of each, rather than
by re-reading the statements an earlier pass had edited. That earlier pass
fixed the two boolean columns it happened to look at and missed the third;
verifying a diff verifies the diff, not the class.

Executed against PostgreSQL 17: PATCH 204, GET 200 with
distraction_free=true and version=1, zero failed requests in the log. A
schema sweep over all three BOOLEAN columns now reports no uncast bind or
read, and no REAL column remains.

docs/verification.md records both as defects 9 and 10 of the live
PostgreSQL run.

+++ migrations/postgres/0004_reading.sql
@@ -102,8 +102,11 @@ CREATE INDEX reader_note_subject ON reader_note (pseud_id, subject_type, subject
 
 CREATE TABLE typography_preference (
     account_id        UUID PRIMARY KEY REFERENCES accounts (id) ON DELETE CASCADE,
-    font_scale        REAL NOT NULL DEFAULT 1.0,
-    line_height       REAL NOT NULL DEFAULT 1.6,
+    -- DOUBLE PRECISION, not REAL: the repository decodes these as f64, and
+    -- SQLite's REAL is already 8 bytes, so REAL (4 bytes) here is the one
+    -- declaration that would not round-trip what the API accepted.
+    font_scale        DOUBLE PRECISION NOT NULL DEFAULT 1.0,
+    line_height       DOUBLE PRECISION NOT NULL DEFAULT 1.6,
     measure           BIGINT NOT NULL DEFAULT 66,
     reader_theme      TEXT NOT NULL DEFAULT 'sepia',
     distraction_free  BOOLEAN NOT NULL DEFAULT FALSE,
